kinesis-data-firehose/dynamic-partitioning/lambda/requirements.txt
aws-cdk.aws-ec2==1.125.0
aws-cdk.aws-iam==1.125.0
aws-cdk.aws-lambda==1.125.0
aws-cdk.aws-logs==1.125.0
aws-cdk.aws-kinesisfirehose==1.125.0
aws-cdk.aws-s3==1.125.0
aws-cdk.core==1.125.0
